requirements.txt
anthropic>=0.40.0
slack-sdk>=3.33.0
python-dotenv>=1.0.0
boto3>=1.35.0
